restructure internals
modify the layout of the project so that not everything is dumped into
src/main.rs. Should make the code more straightforward to extend
